data_selection/baselines/bench/dclm/requirements.txt
attrs
PyYAML
loguru
numpy
pytest
pytest-timeout
nltk
transformers
uniseg
fasttext
jsonlines
blingfire
pandas
sentencepiece
kenlm
langdetect
retrie
https://github.com/kpu/kenlm/archive/master.zip
justext
lxml
cloudpathlib
boto3
moto
# open_lm
scipy
fire
s3fs
braceexpand
pysimdjson
gitpython
Unidecode
beautifulsoup4
zstandard
llm-foundry==0.9.0
lxml_html_clean
